refactor(frontend): Evidence Terminal U2a — extract dashboard sections (#1204)
page.tsx (668 lines) becomes a fetch+enrich+assembly module (~370 lines);
presentational sections move to src/components/dashboard/: helpers.ts (the
5 inline color-map fns, sparkline period parsing, verdict/pipeline style
maps, event date helpers), MarketStrip (the #223 iter-7c IIFE), EventsStrip,
HoldingsSection (incl. HOLDINGS_COLLAPSED_LIMIT), DashboardFooter. Behavior
preserved exactly: all 1472 vitest pass with ZERO test changes; the page
re-exports the seven public helpers so "@/app/page" imports keep working.
One subtle gate preserved deliberately: the footer renders an empty
FreshnessBar when items=[] but details is non-empty (showFreshness prop),
matching the original ?? chain semantics.

Verification: vitest 1472/1472 untouched · tsc clean · build clean ·
responsive matrix 24/24 · full-page screenshot pixel-comparable to before.
Stabilization step before the U2b restructure per docs/UX_REDESIGN_PLAN.md.
